+++ tests/hazmat/primitives/test_cast5.py
@@ -16,7 +16,7 @@ from ...utils import load_nist_vectors
 
 @pytest.mark.supported(
     only_if=lambda backend: backend.cipher_supported(
-        algorithms.CAST5(b"\x00" * 16), modes.ECB()
+        algorithms._CAST5Internal(b"\x00" * 16), modes.ECB()
     ),
     skip_message="Does not support CAST5 ECB",
 )
@@ -25,14 +25,16 @@ class TestCAST5ModeECB:
         load_nist_vectors,
         os.path.join("ciphers", "CAST5"),
         ["cast5-ecb.txt"],
-        lambda key, **kwargs: algorithms.CAST5(binascii.unhexlify((key))),
+        lambda key, **kwargs: algorithms._CAST5Internal(
+            binascii.unhexlify((key))
+        ),
         lambda **kwargs: modes.ECB(),
     )
 
 
 @pytest.mark.supported(
     only_if=lambda backend: backend.cipher_supported(
-        algorithms.CAST5(b"\x00" * 16), modes.CBC(b"\x00" * 8)
+        algorithms._CAST5Internal(b"\x00" * 16), modes.CBC(b"\x00" * 8)
     ),
     skip_message="Does not support CAST5 CBC",
 )
@@ -41,14 +43,16 @@ class TestCAST5ModeCBC:
         load_nist_vectors,
         os.path.join("ciphers", "CAST5"),
         ["cast5-cbc.txt"],
-        lambda key, **kwargs: algorithms.CAST5(binascii.unhexlify((key))),
+        lambda key, **kwargs: algorithms._CAST5Internal(
+            binascii.unhexlify((key))
+        ),
         lambda iv, **kwargs: modes.CBC(binascii.unhexlify(iv)),
     )
 
 
 @pytest.mark.supported(
     only_if=lambda backend: backend.cipher_supported(
-        algorithms.CAST5(b"\x00" * 16), modes.OFB(b"\x00" * 8)
+        algorithms._CAST5Internal(b"\x00" * 16), modes.OFB(b"\x00" * 8)
     ),
     skip_message="Does not support CAST5 OFB",
 )
@@ -57,14 +61,16 @@ class TestCAST5ModeOFB:
         load_nist_vectors,
         os.path.join("ciphers", "CAST5"),
         ["cast5-ofb.txt"],
-        lambda key, **kwargs: algorithms.CAST5(binascii.unhexlify((key))),
+        lambda key, **kwargs: algorithms._CAST5Internal(
+            binascii.unhexlify((key))
+        ),
         lambda iv, **kwargs: modes.OFB(binascii.unhexlify(iv)),
     )
 
 
 @pytest.mark.supported(
     only_if=lambda backend: backend.cipher_supported(
-        algorithms.CAST5(b"\x00" * 16), modes.CFB(b"\x00" * 8)
+        algorithms._CAST5Internal(b"\x00" * 16), modes.CFB(b"\x00" * 8)
     ),
     skip_message="Does not support CAST5 CFB",
 )
@@ -73,6 +79,8 @@ class TestCAST5ModeCFB:
         load_nist_vectors,
         os.path.join("ciphers", "CAST5"),
         ["cast5-cfb.txt"],
-        lambda key, **kwargs: algorithms.CAST5(binascii.unhexlify((key))),
+        lambda key, **kwargs: algorithms._CAST5Internal(
+            binascii.unhexlify((key))
+        ),
         lambda iv, **kwargs: modes.CFB(binascii.unhexlify(iv)),
     )