+++ tests/hazmat/primitives/test_seed.py
@@ -16,7 +16,7 @@ from ...utils import load_nist_vectors
 
 @pytest.mark.supported(
     only_if=lambda backend: backend.cipher_supported(
-        algorithms.SEED(b"\x00" * 16), modes.ECB()
+        algorithms._SEEDInternal(b"\x00" * 16), modes.ECB()
     ),
     skip_message="Does not support SEED ECB",
 )
@@ -25,14 +25,16 @@ class TestSEEDModeECB:
         load_nist_vectors,
         os.path.join("ciphers", "SEED"),
         ["rfc-4269.txt"],
-        lambda key, **kwargs: algorithms.SEED(binascii.unhexlify((key))),
+        lambda key, **kwargs: algorithms._SEEDInternal(
+            binascii.unhexlify((key))
+        ),
         lambda **kwargs: modes.ECB(),
     )
 
 
 @pytest.mark.supported(
     only_if=lambda backend: backend.cipher_supported(
-        algorithms.SEED(b"\x00" * 16), modes.CBC(b"\x00" * 16)
+        algorithms._SEEDInternal(b"\x00" * 16), modes.CBC(b"\x00" * 16)
     ),
     skip_message="Does not support SEED CBC",
 )
@@ -41,14 +43,16 @@ class TestSEEDModeCBC:
         load_nist_vectors,
         os.path.join("ciphers", "SEED"),
         ["rfc-4196.txt"],
-        lambda key, **kwargs: algorithms.SEED(binascii.unhexlify((key))),
+        lambda key, **kwargs: algorithms._SEEDInternal(
+            binascii.unhexlify((key))
+        ),
         lambda iv, **kwargs: modes.CBC(binascii.unhexlify(iv)),
     )
 
 
 @pytest.mark.supported(
     only_if=lambda backend: backend.cipher_supported(
-        algorithms.SEED(b"\x00" * 16), modes.OFB(b"\x00" * 16)
+        algorithms._SEEDInternal(b"\x00" * 16), modes.OFB(b"\x00" * 16)
     ),
     skip_message="Does not support SEED OFB",
 )
@@ -57,14 +61,16 @@ class TestSEEDModeOFB:
         load_nist_vectors,
         os.path.join("ciphers", "SEED"),
         ["seed-ofb.txt"],
-        lambda key, **kwargs: algorithms.SEED(binascii.unhexlify((key))),
+        lambda key, **kwargs: algorithms._SEEDInternal(
+            binascii.unhexlify((key))
+        ),
         lambda iv, **kwargs: modes.OFB(binascii.unhexlify(iv)),
     )
 
 
 @pytest.mark.supported(
     only_if=lambda backend: backend.cipher_supported(
-        algorithms.SEED(b"\x00" * 16), modes.CFB(b"\x00" * 16)
+        algorithms._SEEDInternal(b"\x00" * 16), modes.CFB(b"\x00" * 16)
     ),
     skip_message="Does not support SEED CFB",
 )
@@ -73,6 +79,8 @@ class TestSEEDModeCFB:
         load_nist_vectors,
         os.path.join("ciphers", "SEED"),
         ["seed-cfb.txt"],
-        lambda key, **kwargs: algorithms.SEED(binascii.unhexlify((key))),
+        lambda key, **kwargs: algorithms._SEEDInternal(
+            binascii.unhexlify((key))
+        ),
         lambda iv, **kwargs: modes.CFB(binascii.unhexlify(iv)),
     )